Working through LYSE
Third try is the charm

Had a good time working through 'sum' and 'prod' are an exercise for the reader.
That was moderately unexpected, and just non-trivial enough to
be a proper challenge. First attempt was to use recursion, which
did not go well (no surprise, I expect). Second attempt was to
do the sum piece by piece in rpn/2 and to write the sum value
back to the list at the front or end of the list, which did not
work.

Thinking it through, logically the correct operation here is a foldl
through the remainder of the list. Just need to remember what is the
identity element for these two operations. Makes me long for the
good old days of studying rings back in ~2003-2004.